Added the default value for the slave-to-master part of WB sub-buses. The default value asserts ERR signal. If the user does not connect a sub-bus in the user logic, an attempt to access it returns the WB error.
Current implementation places the "c_WB_SLAVE_OUT_ERR" definition
in each generated package. That is not the most elegant solution,
but is not a huge problem, as each block uses only its
own package (and does not uses packages of its slaves).
Another solution would be to put the "c_WB_SLAVE_OUT_ERR" definition
into the "extension" of the general-cores wishbone_pkg and require
using it in all generated agwb blocks (to be considered).