Revert "Added the new optional "registered" attribute to the block definition."
I dislike this solution, as the decision must be done per-block, not per-instance.
Probably the user may prefer to enforce using the registered crossbar in certain instances.
Therefore I'll probably change the implementation, so that the user may specify the "registered" mode in the user logic, when instantiating the AGWB-generated block.

This reverts commit d61bee1cf95b6e3f529a2fe1eeab500a22a33d10.