Added the "trigger" attribute for bitfields, to avoid the problem found by @wfjm (see below).
Setting `trigger="1"` marks the bitfield as a trigger.
The bitfield anded with register's stb may be used to trigger the action.
The bitfield is zeroed when the register is read, so it is safe to
use it in RMW operations.

The problem described by @wfjm occured in the situation, when
the control register contained two bitfields (let's call them F0 and F1)
used to trigger the action.
When somebody sets F0 to one, it triggers the action, but the bitfield
remains set in the underlying register.
The further setting of F1 reads the register, setd F1 by OR and writes
it back. That results in triggering both actions at the same time.
Masking the "trigger" fields at readout solves the problem.